Seperate game image into component
