+++ src/ui/components/mod.rs
@@ -2,3 +2,5 @@ mod steam_user_select;
 mod game_image_button;
 
 pub use steam_user_select::render_user_select;
+pub use game_image_button::render_image_from_path;
+pub use game_image_button::render_image_from_path_or_url;
